+++ graysvr/CCharSkill.cpp
@@ -4033,7 +4033,7 @@ bool CChar::Skill_Start( SKILL_TYPE skill, int iDifficulty )
 	{
 		if ( skill != SKILL_NONE && !IsSkillBase(skill) && !IsSkillNPC(skill) )
 		{
-			DEBUG_ERR(("UID:0%x Bad Skill %d for '%s'\n", GetUID(), skill, GetName()));
+			DEBUG_ERR(("UID:0%x Bad Skill %d for '%s'\n", (DWORD)GetUID(), skill, GetName()));
 			return false;
 		}
 		m_Act_SkillCurrent = skill;
